// src/shared/api/generated/models/AccountProfileResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CRM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccountProfileResponse
 */
export interface AccountProfileResponse {
    /**
     * Логин пользователя
     * @type {string}
     * @memberof AccountProfileResponse
     */
    login: string;
    /**
     * 
     * @type {string}
     * @memberof AccountProfileResponse
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof AccountProfileResponse
     */
    lastName: string;
    /**
     * Дата рождения пользователя (от 16 до 100 лет)
     * @type {Date}
     * @memberof AccountProfileResponse
     */
    birthDay?: Date;
    /**
     * 
     * @type {string}
     * @memberof AccountProfileResponse
     */
    avatarUrl?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AccountProfileResponse
     */
    isPrivate?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof AccountProfileResponse
     */
    isAdmin?: boolean;
}

/**
 * Check if a given object implements the AccountProfileResponse interface.
 */
export function instanceOfAccountProfileResponse(value: object): value is AccountProfileResponse {
    if (!('login' in value) || value['login'] === undefined) return false;
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    return true;
}

export function AccountProfileResponseFromJSON(json: any): AccountProfileResponse {
    return AccountProfileResponseFromJSONTyped(json, false);
}

export function AccountProfileResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccountProfileResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'login': json['login'],
        'firstName': json['first_name'],
        'lastName': json['last_name'],
        'birthDay': json['birth_day'] == null ? undefined : (new Date(json['birth_day'])),
        'avatarUrl': json['avatar_url'] == null ? undefined : json['avatar_url'],
        'isPrivate': json['is_private'] == null ? undefined : json['is_private'],
        'isAdmin': json['is_admin'] == null ? undefined : json['is_admin'],
    };
}

export function AccountProfileResponseToJSON(json: any): AccountProfileResponse {
    return AccountProfileResponseToJSONTyped(json, false);
}

export function AccountProfileResponseToJSONTyped(value?: AccountProfileResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'login': value['login'],
        'first_name': value['firstName'],
        'last_name': value['lastName'],
        'birth_day': value['birthDay'] == null ? undefined : ((value['birthDay']).toISOString().substring(0,10)),
        'avatar_url': value['avatarUrl'],
        'is_private': value['isPrivate'],
        'is_admin': value['isAdmin'],
    };
}